requirements.txt
crewai
crewai_tools
load_dotenv
langchain_google_genai
